+++ src/single_threaded.rs
@@ -855,6 +855,105 @@ impl<'a, K: Key, S> IntoIterator for &'a Rodeo<K, S> {
     }
 }
 
+compile! {
+    if #[feature = "serialize"] {
+        use core::num::NonZeroUsize;
+        use serde::{
+            de::{Deserialize, Deserializer},
+            ser::{Serialize, Serializer},
+        };
+    }
+}
+
+#[cfg(feature = "serialize")]
+impl<K, H> Serialize for Rodeo<K, H> {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        // Serialize all of self as a `Vec<String>`
+        self.strings.serialize(serializer)
+    }
+}
+
+#[cfg(feature = "serialize")]
+impl<'de, K: Key, S: BuildHasher + Default> Deserialize<'de> for Rodeo<K, S> {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let vector: Vec<String> = Vec::deserialize(deserializer)?;
+        let capacity = {
+            let total_bytes = vector.iter().map(|s| s.len()).sum::<usize>();
+            let total_bytes =
+                NonZeroUsize::new(total_bytes).unwrap_or_else(|| Capacity::default().bytes());
+
+            Capacity::new(vector.len(), total_bytes)
+        };
+
+        let hasher: S = Default::default();
+        let mut strings = Vec::with_capacity(capacity.strings);
+        let mut map = HashMap::with_capacity_and_hasher(capacity.strings, ());
+        let mut arena = Arena::new(capacity.bytes, capacity.bytes.get() * 2);
+
+        for (key, string) in vector.into_iter().enumerate() {
+            let allocated = unsafe {
+                arena
+                    .store_str(&string)
+                    .expect("failed to allocate enough memory")
+            };
+
+            let hash = {
+                let mut state = hasher.build_hasher();
+                allocated.hash(&mut state);
+
+                state.finish()
+            };
+
+            // Get the map's entry that the string should occupy
+            let entry = map.raw_entry_mut().from_hash(hash, |key: &K| {
+                // Safety: The index given by `key` will be in bounds of the strings vector
+                let key_string: &str = unsafe { index_unchecked!(strings, key.into_usize()) };
+
+                // Compare the requested string against the key's string
+                allocated == key_string
+            });
+
+            match entry {
+                RawEntryMut::Occupied(..) => {
+                    debug_assert!(false, "re-interned a key while deserializing");
+                }
+                RawEntryMut::Vacant(entry) => {
+                    // Create the key from the vec's index that the string will hold
+                    let key =
+                        K::try_from_usize(key).expect("failed to create key while deserializing");
+
+                    // Push the allocated string to the strings vector
+                    strings.push(allocated);
+
+                    // Insert the key with the hash of the string that it points to, reusing the hash we made earlier
+                    entry.insert_with_hasher(hash, key, (), |key| {
+                        let key_string: &str =
+                            unsafe { index_unchecked!(strings, key.into_usize()) };
+
+                        let mut state = hasher.build_hasher();
+                        key_string.hash(&mut state);
+
+                        state.finish()
+                    });
+                }
+            }
+        }
+
+        Ok(Self {
+            map,
+            hasher,
+            strings,
+            arena,
+        })
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use crate::{hasher::RandomState, Capacity, Key, MemoryLimits, MicroSpur, Rodeo, Spur};
@@ -1344,6 +1443,52 @@ mod tests {
         ) {
             assert_eq!(key, Spur::try_from_usize(expected_key).unwrap());
             assert_eq!(string, expected_string);
+        }
+    }
+
+    #[test]
+    #[cfg(feature = "serialize")]
+    fn empty_serialize() {
+        let rodeo = Rodeo::default();
+
+        let ser = serde_json::to_string(&rodeo).unwrap();
+        let ser2 = serde_json::to_string(&rodeo).unwrap();
+        assert_eq!(ser, ser2);
+
+        let deser: Rodeo = serde_json::from_str(&ser).unwrap();
+        assert!(deser.is_empty());
+        let deser2: Rodeo = serde_json::from_str(&ser2).unwrap();
+        assert!(deser2.is_empty());
+    }
+
+    #[test]
+    #[cfg(feature = "serialize")]
+    fn filled_serialize() {
+        let mut rodeo = Rodeo::default();
+        let a = rodeo.get_or_intern("a");
+        let b = rodeo.get_or_intern("b");
+        let c = rodeo.get_or_intern("c");
+        let d = rodeo.get_or_intern("d");
+
+        let ser = serde_json::to_string(&rodeo).unwrap();
+        let ser2 = serde_json::to_string(&rodeo).unwrap();
+        assert_eq!(ser, ser2);
+
+        let deser: Rodeo = serde_json::from_str(&ser).unwrap();
+        let deser2: Rodeo = serde_json::from_str(&ser2).unwrap();
+
+        for (((correct_key, correct_str), (key1, str1)), (key2, str2)) in
+            [(a, "a"), (b, "b"), (c, "c"), (d, "d")]
+                .iter()
+                .copied()
+                .zip(&deser)
+                .zip(&deser2)
+        {
+            assert_eq!(correct_key, key1);
+            assert_eq!(correct_key, key2);
+
+            assert_eq!(correct_str, str1);
+            assert_eq!(correct_str, str2);
         }
     }
 }